% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coordinate_uncertainty.R
\name{add_coordinate_uncertainty}
\alias{add_coordinate_uncertainty}
\title{Add coordinate uncertainty to observations}
\usage{
add_coordinate_uncertainty(observations, coords_uncertainty_meters = 25)
}
\arguments{
\item{observations}{An sf object with POINT geometry or a simple
dataframe representing the observations. This object contains the observation
points to which the coordinate uncertainty will be added.}

\item{coords_uncertainty_meters}{A numeric value or a vector of numeric
values representing the coordinate uncertainty (in meters) associated with
each observation. If a single numeric value is provided, it will be applied
to all observations. If a numeric vector is provided, it must be the same
length as the number of observations.}
}
\value{
The input data frame or an sf object with POINT geometry, with an
additional column named \code{coordinateUncertaintyInMeters} that contains the
coordinate uncertainty values in meters.
}
\description{
This function adds a column to the input dataframe or sf object containing
the coordinate uncertainty for each observation, measured in meters.
}
\examples{
# Create dataframe with sampling status column
observations_data <- data.frame(
    time_point = 1,
    sampling_prob = seq(0.5, 1, 0.1)
  )

# provide a fixed uncertainty for all points
add_coordinate_uncertainty(
  observations_data,
  coords_uncertainty_meters = 1000
 )

# add variability in uncertainty. For example, using gamma distribution
uncertainty_vec <- seq(50, 100, 10)

add_coordinate_uncertainty(
  observations_data,
  coords_uncertainty_meters = uncertainty_vec
)
}
\seealso{
Other main: 
\code{\link{filter_observations}()},
\code{\link{grid_designation}()},
\code{\link{sample_observations}()},
\code{\link{simulate_occurrences}()},
\code{\link{virtualsample_to_sf}()}
}
\concept{main}
