% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_observations.R
\name{filter_observations}
\alias{filter_observations}
\title{Filter detected observations}
\usage{
filter_observations(observations_total, invert = FALSE)
}
\arguments{
\item{observations_total}{An sf object with POINT geometry or a simple
dataframe with \code{sampling_status} column containing values \code{"detected"}.
This format is created by \code{sample_observations()}.}

\item{invert}{Logical. If \code{FALSE} (default), filter \code{"detected"}
occurrences. Otherwise, filter all other occurrences.}
}
\value{
A dataframe or an sf object with POINT geometry containing detected
occurrences (if \code{invert = FALSE}), or other occurrences (if \code{invert = TRUE}).
}
\description{
The function filters observations from all observations based on a
\code{sampling_status} column, e.g. created by \code{sample_observations()}.
}
\examples{
# Load packages
library(sf)
library(dplyr)

# Set seed for reproducibility
set.seed(123)

# Simulate some occurrence data with coordinates and time points
num_points <- 10
occurrences <- data.frame(
  lon = runif(num_points, min = -180, max = 180),
  lat = runif(num_points, min = -90, max = 90),
  time_point = 0
  )

# Convert the occurrence data to an sf object
occurrences_sf <- st_as_sf(occurrences, coords = c("lon", "lat"))

# Sample observations without sampling bias
observations_total_sf <- sample_observations(
  occurrences_sf,
  detection_probability = 0.8,
  sampling_bias = "no_bias",
  seed = 123
  )

# Filter detected observations
filter_observations(observations_total_sf)

# Filter undetected observations
filter_observations(observations_total_sf, invert = TRUE)
}
\seealso{
Other main: 
\code{\link{add_coordinate_uncertainty}()},
\code{\link{grid_designation}()},
\code{\link{sample_observations}()},
\code{\link{simulate_occurrences}()}
}
\concept{main}
