% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxonomy.R
\name{generate_taxonomy}
\alias{generate_taxonomy}
\title{Generate a taxonomic hierarchy}
\usage{
generate_taxonomy(
  num_species,
  num_genera,
  num_families,
  num_orders = 1,
  num_classes = 1,
  num_phyla = 1,
  num_kingdoms = 1,
  seed = NA
)
}
\arguments{
\item{num_species}{Number of species to generate, or a dataframe. With a
dataframe, the function will create a species with taxonomic hierarchy for
each row. The original columns of the dataframe will be retained in the
output.}

\item{num_genera}{Number of genera to generate.}

\item{num_families}{Number of families to generate.}

\item{num_orders}{Number of orders to generate. Defaults to 1.}

\item{num_classes}{Number of classes to generate. Defaults to 1.}

\item{num_phyla}{Number of phyla to generate. Defaults to 1.}

\item{num_kingdoms}{Number of kingdoms to generate. Defaults to 1.}

\item{seed}{The seed for random number generation to make results
reproducible. If \code{NA} (the default), no seed is used.}
}
\value{
A data frame with the taxonomic classification of each species. If
\code{num_species} is a dataframe, the taxonomic classification is added to this
input dataframe. The original columns of the dataframe will be retained in
the output.
}
\description{
This function generates a random taxonomic hierarchy for a specified numbers
of species, genera, families, orders, classes, phyla, and kingdoms. The
output is a data frame with the hierarchical classification for each species.
}
\details{
The function works by randomly assigning species to genera, genera to
families, families to orders, orders to classes, classes to phyla, and phyla
to kingdoms. Sampling is done with replacement, meaning that multiple
lower-level taxa (e.g., species) can be assigned to the same higher-level
taxon (e.g., genus).
}
\examples{
# Create simple taxonomic hierarchy
generate_taxonomy(
  num_species = 5,
  num_genera = 3,
  num_families = 2,
  seed = 123)

# Add taxonomic hierarchy to a dataframe
existing_df <- data.frame(
  count = c(1, 2, 5, 4, 8, 9, 3),
  det_prob = c(0.9, 0.9, 0.9, 0.8, 0.5, 0.2, 0.2)
  )

generate_taxonomy(
  num_species = existing_df,
  num_genera = 4,
  num_families = 2,
  seed = 125)
}
\seealso{
Other multispecies: 
\code{\link{map_add_coordinate_uncertainty}()},
\code{\link{map_filter_observations}()},
\code{\link{map_grid_designation}()},
\code{\link{map_sample_observations}()},
\code{\link{map_simulate_occurrences}()}
}
\concept{multispecies}
