% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_designation.R
\name{grid_designation}
\alias{grid_designation}
\title{Observations to grid designation to create a data cube}
\usage{
grid_designation(
  observations,
  grid,
  id_col = "row_names",
  seed = NA,
  aggregate = TRUE,
  randomisation = c("uniform", "normal"),
  p_norm = ifelse(tolower(randomisation[1]) == "uniform", NA, 0.95)
)
}
\arguments{
\item{observations}{An sf object with POINT geometry and a \code{time_point} and
\code{coordinateUncertaintyInMeters} column. If the latter column is not present,
the function will assume no uncertainty (zero meters) around the observation
points.}

\item{grid}{An sf object with POLYGON geometry (usually a grid) to which
observations should be designated.}

\item{id_col}{The column name containing unique IDs for each grid cell. If
\code{"row_names"} (the default), a new column \code{id} is created where the row names
represent the unique IDs.}

\item{seed}{A positive numeric value setting the seed for random number
generation to ensure reproducibility. If \code{NA} (default), then \code{set.seed()}
is not called at all. If not \code{NA}, then the random number generator state is
reset (to the state before calling this function) upon exiting this function.}

\item{aggregate}{Logical. If \code{TRUE} (default), returns data cube in
aggregated form (grid with the number of observations per grid cell).
Otherwise, returns sampled points within the uncertainty circle.}

\item{randomisation}{Character. Method used for sampling within the
uncertainty circle around each observation. \code{"uniform"} (default) means each
point in the uncertainty circle has an equal probability of being selected.
The other option is \code{"normal"}, where a point is sampled from a bivariate
Normal distribution with means equal to the observation point and variance
such that \code{p_norm} \% of all possible samples from this Normal distribution
fall within the uncertainty circle.
See \code{sample_from_binormal_circle()}.}

\item{p_norm}{A numeric value between 0 and 1, used only if
\code{randomisation = "normal"}. The proportion of all possible samples from a
bivariate Normal distribution that fall within the uncertainty circle.
Default is 0.95.}
}
\value{
If \code{aggregate = TRUE}, an sf object with POLYGON geometry
containing the grid cells, an \code{n} column with the number of observations per
grid cell, and a \code{min_coord_uncertainty} column with the minimum coordinate
uncertainty per grid cell. If \code{aggregate = FALSE}, an sf object with POINT
geometry containing the sampled observations within the uncertainty circles,
and a \code{coordinateUncertaintyInMeters} column with the coordinate uncertainty
for each observation.
}
\description{
This function designates observations to cells of a given grid to create an
aggregated data cube.
}
\examples{
library(sf)
library(dplyr)

set.seed(123)

# Create four random points
n_points <- 4
xlim <- c(3841000, 3842000)
ylim <- c(3110000, 3112000)
coordinate_uncertainty <- rgamma(n_points, shape = 5, rate = 0.1)

observations_sf <- data.frame(
  lat = runif(n_points, ylim[1], ylim[2]),
  long = runif(n_points, xlim[1], xlim[2]),
  time_point = 1,
  coordinateUncertaintyInMeters = coordinate_uncertainty
) \%>\%
  st_as_sf(coords = c("long", "lat"), crs = 3035)

# Add buffer uncertainty in meters around points
observations_buffered <- observations_sf \%>\%
  st_buffer(observations_sf$coordinateUncertaintyInMeters)

# Create grid
grid_df <- st_make_grid(
  observations_buffered,
  square = TRUE,
  cellsize = c(200, 200)
) \%>\%
  st_sf()

# Create occurrence cube
grid_designation(
  observations = observations_sf,
  grid = grid_df,
  seed = 123
)
}
\seealso{
Other main: 
\code{\link{add_coordinate_uncertainty}()},
\code{\link{filter_observations}()},
\code{\link{sample_observations}()},
\code{\link{simulate_occurrences}()}
}
\concept{main}
