% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_filter_observations.R
\name{map_filter_observations}
\alias{map_filter_observations}
\title{Map \code{filter_observations()} function over multiple species}
\usage{
map_filter_observations(df, nested = TRUE, arg_list = NA)
}
\arguments{
\item{df}{A dataframe containing multiple rows. Each row is considered a
different species. The columns are function arguments with values used for
mapping \code{filter_observations()} for each species. \code{df} can have columns that
are not used by this function. They will be retained in the output.}

\item{nested}{Logical. If \code{TRUE} (default), retain list-column containing
sf objects/dataframes calculated by \code{filter_observations()}. Otherwise,
expand this list-column into rows and columns.}

\item{arg_list}{A named list or \code{NA}. If \code{NA} (default), the function assumes
column names in \code{df} are identical to argument names of
\code{filter_observations()}. If column names are not identical, they need to be
specified as a named list where the names are the argument names of
\code{filter_observations()}.}
}
\value{
In case of \code{nested = TRUE}, a dataframe identical to the input
dataframe \code{df}, but with an extra list-column called \code{observations}
containing an sf object with POINT geometry or simple dataframe for each row
computed by \code{filter_observations()}. In case of \code{nested = FALSE}, this
list-column is expanded into additional rows and columns.
}
\description{
The function executes \code{filter_observations()} over multiple rows of a
dataframe, representing multiple different species, containing potentially
different function arguments over multiple columns.
}
\examples{
# Load packages
library(sf)
library(dplyr)

# Create polygon
plgn <- st_polygon(list(cbind(c(5, 10, 8, 2, 3, 5), c(2, 1, 7, 9, 5, 2))))

## Example with simple column names
# Specify dataframe for 3 species with custom function arguments
species_dataset_df <- tibble(
  taxonID = c("species1", "species2", "species3"),
  plgn = rep(list(plgn), 3),
  initial_average_abundance = c(50, 100, 500),
  n_time_points = rep(6, 3),
  temporal_function = c(simulate_random_walk, simulate_random_walk, NA),
  sd_step = c(1, 1, NA),
  spatial_autocorr = "random",
  detection_probability = c(0.8, 0.9, 1),
  invert = FALSE,
  seed = 123)

# Simulate occurrences
sim_occ1 <- map_simulate_occurrences(df = species_dataset_df)

# Sample observations
samp_obs1 <- map_sample_observations(df = sim_occ1)

# Filter observations
filter_obs_nested <- map_filter_observations(df = samp_obs1)
filter_obs_nested

# Unnest output and create sf object again
filter_obs_unnested <- map_filter_observations(df = samp_obs1,
                                               nested = FALSE)
filter_obs_unnested \%>\%
   st_sf()


## Example with deviating column names
# Specify dataframe for 3 species with custom function arguments
species_dataset_df2 <- species_dataset_df \%>\%
  rename(polygon = plgn,
         sd = sd_step,
         det_prob = detection_probability,
         inv = invert)

# Create named list for argument conversion
arg_conv_list <- list(
    plgn = "polygon",
    sd_step = "sd",
    detection_probability = "det_prob",
    invert = "inv"
  )

# Simulate occurrences
sim_occ2 <- map_simulate_occurrences(
  df = species_dataset_df2,
  arg_list = arg_conv_list)

# Sample observations
samp_obs2 <- map_sample_observations(
  df = sim_occ2,
  arg_list = arg_conv_list)

# Filter observations
map_filter_observations(
  df = samp_obs2,
  arg_list = arg_conv_list)
}
\seealso{
Other multispecies: 
\code{\link{map_sample_observations}()},
\code{\link{map_simulate_occurrences}()}
}
\concept{multispecies}
