% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_uniform_circle.R
\name{sample_from_uniform_circle}
\alias{sample_from_uniform_circle}
\title{Sample from a circle using the Uniform distribution}
\usage{
sample_from_uniform_circle(observations, missing_uncertainty = 1000, seed = NA)
}
\arguments{
\item{observations}{An sf object with POINT geometry and a \code{time_point} and
\code{coordinateUncertaintyInMeters} column. If the former column is not present,
the function will assume a single time point. If the latter column is not
present, the function will assume no uncertainty (zero meters) around the
observation points.}

\item{missing_uncertainty}{A positive numeric value (default: 1000 m) used to
replace missing (\code{NA}) values in the \code{coordinateUncertaintyInMeters} column.
This ensures that all observations have a defined uncertainty radius for
sampling. Only applied when the column is present but contains \code{NA} values;
if the column itself is absent, a value of 0 is assumed instead.}

\item{seed}{A positive numeric value setting the seed for random number
generation to ensure reproducibility. If \code{NA} (default), then \code{set.seed()}
is not called at all. If not \code{NA}, then the random number generator state is
reset (to the state before calling this function) upon exiting this function.}
}
\value{
An sf object with POINT geometry containing the locations of the
sampled occurrences and a \code{coordinateUncertaintyInMeters} column containing
the coordinate uncertainty for each observation.
}
\description{
This function samples a new observations point of a species within the
uncertainty circle around each observation assuming a Uniform distribution.
}
\examples{
library(sf)
library(dplyr)

# Create four random points
n_points <- 4
xlim <- c(3841000, 3842000)
ylim <- c(3110000, 3112000)
coordinate_uncertainty <- rgamma(n_points, shape = 5, rate = 0.1)

observations_sf <- data.frame(
  lat = runif(n_points, ylim[1], ylim[2]),
  long = runif(n_points, xlim[1], xlim[2]),
  time_point = 1,
  coordinateUncertaintyInMeters = coordinate_uncertainty
) \%>\%
  st_as_sf(coords = c("long", "lat"), crs = 3035)

# Sample points within uncertainty circles according to uniform rules
sample_from_uniform_circle(
  observations = observations_sf,
  seed = 123
)
}
\seealso{
Other designation: 
\code{\link{sample_from_binormal_circle}()}
}
\concept{designation}
