% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_observations.R
\name{sample_observations}
\alias{sample_observations}
\title{Sample observations from a larger occurrence dataset}
\usage{
sample_observations(
  occurrences,
  detection_probability = 1,
  sampling_bias = c("no_bias", "polygon", "manual"),
  bias_area = NA,
  bias_strength = 1,
  bias_weights = NA,
  seed = NA
)
}
\arguments{
\item{occurrences}{An sf object with POINT geometry representing the
occurrences.}

\item{detection_probability}{A numeric value between 0 and 1 representing the
probability of detecting the species.}

\item{sampling_bias}{A character string specifying the method to generate a
sampling bias.
Options are \code{"no_bias"}, \code{"polygon"}, or \code{"manual"}.
\describe{
\item{\code{"no_bias"}}{No bias is applied (default).}
\item{\code{"polygon"}}{Bias the sampling within a polygon. Provide the polygon
to \code{bias_area} and the bias strength to \code{bias_strength}.}
\item{\code{"manual"}}{Bias the sampling manually using a grid. Provide the grid
layer in which each cell contains the probability of being sampled to
\code{bias_weights}.}
}}

\item{bias_area}{An \code{sf} object with POLYGON geometry, or \code{NA}. Only used if
\code{sampling_bias = "polygon"}. This defines the area in which the sampling will
be biased.}

\item{bias_strength}{A positive numeric value, or \code{NA}. Only used if
\code{sampling_bias = "polygon"}. The value represents the strength of
the bias to be applied within the \code{bias_area}. Values greater than 1 will
increase the sampling probability within the polygon relative to outside
(oversampling), while values between 0 and 1 will decrease it
(undersampling). For instance, a value of 50 will make the probability 50
times higher within the \code{bias_area} compared to outside, whereas a value of
0.5 will make it half as likely.}

\item{bias_weights}{A grid layer (an sf object with POLYGON geometry), or
\code{NA}. Only used if \code{sampling_bias = "manual"}. The grid of bias weights to be
applied. This sf object should contain a \code{bias_weight} column with the
weights per grid cell. Higher weights increase the probability of sampling.
Weights can be numeric values between 0 and 1 or positive integers, which
will be rescaled to values between 0 and 1.}

\item{seed}{A positive numeric value setting the seed for random number
generation to ensure reproducibility. If \code{NA} (default), then \code{set.seed()}
is not called at all. If not \code{NA}, then the random number generator state is
reset (to the state before calling this function) upon exiting this function.}
}
\value{
An sf object with POINT geometry containing the locations of the
occurrence with detection status. The object includes the following columns:
\describe{
\item{\code{detection_probability}}{The detection probability for each
occurrence (will be the same for all).}
\item{\code{bias_weight}}{The sampling probability based on sampling bias for
each occurrence.}
\item{\code{sampling_probability}}{The combined sampling probability from
detection probability and sampling bias for each occurrence.}
\item{\code{sampling_status}}{Indicates whether the occurrence was detected
(\code{"detected"}) or not (\code{"undetected"}). Detected occurrences are called
observations.}
}
}
\description{
The function computes observations from occurrences based on detection
probability and sampling bias by implementing a Bernoulli trial.
}
\examples{
# Load packages
library(sf)
library(dplyr)

# Set seed for reproducibility
set.seed(123)

# Simulate some occurrence data with coordinates and time points
num_points <- 10
occurrences <- data.frame(
  lon = runif(num_points, min = -180, max = 180),
  lat = runif(num_points, min = -90, max = 90),
  time_point = 0
  )

# Convert the occurrence data to an sf object
occurrences_sf <- st_as_sf(occurrences, coords = c("lon", "lat"))

# 1. Sample observations without sampling bias
sample_observations(
  occurrences_sf,
  detection_probability = 0.8,
  sampling_bias = "no_bias",
  seed = 123
  )

# 2. Sample observations with sampling bias in a polygon
# Create bias_area polygon overlapping two of the points
selected_observations <- st_union(occurrences_sf[2:3,])
bias_area <- st_convex_hull(selected_observations) \%>\%
  st_buffer(dist = 50) \%>\%
  st_as_sf()

sample_observations(
  occurrences_sf,
  detection_probability = 0.8,
  sampling_bias = "polygon",
  bias_area = bias_area,
  bias_strength = 2,
  seed = 123
  )

# 3. Sample observations with sampling bias given manually in a grid
# Create raster grid with bias weights between 0 and 1
grid <- st_make_grid(occurrences_sf) \%>\%
  st_sf() \%>\%
  mutate(bias_weight = runif(n(), min = 0, max = 1))

sample_observations(
  occurrences_sf,
  detection_probability = 0.8,
  sampling_bias = "manual",
  bias_weights = grid,
  seed = 123
  )
}
\seealso{
Other main: 
\code{\link{add_coordinate_uncertainty}()},
\code{\link{filter_observations}()},
\code{\link{grid_designation}()},
\code{\link{simulate_occurrences}()}
}
\concept{main}
