% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_occurrences_from_raster.R
\name{sample_occurrences_from_raster}
\alias{sample_occurrences_from_raster}
\title{Sample occurrences from spatial random field}
\usage{
sample_occurrences_from_raster(raster, time_series, seed = NA)
}
\arguments{
\item{raster}{A SpatRaster object (see \code{\link[terra:rast]{terra::rast()}}).}

\item{time_series}{A vector with the number of occurrences per time point.}

\item{seed}{A positive numeric value setting the seed for random number
generation to ensure reproducibility. If \code{NA} (default), no seed is used.}
}
\value{
An sf object with POINT geometry containing the locations of the
simulated occurrences, a \code{time_point} column indicating the associated
time point for each occurrence and columns used as weights for sampling.
If the raster is created with \code{create_spatial_pattern()}, the column
\code{sampling_p1} is used.
}
\description{
This function draws point occurrences from a spatial random field represented
by a raster. Points are sampled based on the values in the raster, with the
number of occurrences specified for each time step.
}
\examples{
# Load packages
library(sf)
library(ggplot2)
library(tidyterra)

# Create polygon
plgn <- st_polygon(list(cbind(c(5, 10, 8, 2, 3, 5), c(2, 1, 7, 9, 5, 2))))

## Medium scale clustering
# Create the random field
rs_pattern_clustered <- create_spatial_pattern(
  polygon = plgn,
  resolution = 0.1,
  spatial_pattern = "clustered",
  seed = 123)

# Sample 200 occurrences from random field
pts_occ_clustered <- sample_occurrences_from_raster(
  raster = rs_pattern_clustered,
  time_series = 200,
  seed = 123)

ggplot() +
  geom_spatraster(data = rs_pattern_clustered) +
  geom_sf(data = pts_occ_clustered) +
  scale_fill_continuous(type = "viridis") +
  theme_minimal()

## Large scale clustering
# Create the random field
rs_pattern_large <- create_spatial_pattern(
  polygon = plgn,
  resolution = 0.1,
  spatial_pattern = 100,
  seed = 123)

# Sample 200 occurrences from random field
pts_occ_large <- sample_occurrences_from_raster(
  raster = rs_pattern_large,
  time_series = 200,
  seed = 123)

ggplot() +
  geom_spatraster(data = rs_pattern_large) +
  geom_sf(data = pts_occ_large) +
  scale_fill_continuous(type = "viridis") +
  theme_minimal()
}
\seealso{
Other occurrence: 
\code{\link{create_spatial_pattern}()},
\code{\link{simulate_random_walk}()},
\code{\link{simulate_timeseries}()}
}
\concept{occurrence}
