% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_random_walk.R
\name{simulate_random_walk}
\alias{simulate_random_walk}
\title{Simulate a random walk over time}
\usage{
simulate_random_walk(
  initial_average_occurrences = 50,
  n_time_points = 10,
  sd_step = 0.05,
  seed = NA
)
}
\arguments{
\item{initial_average_occurrences}{A positive numeric value indicating the
average number of occurrences to be simulated at the first time point.}

\item{n_time_points}{A positive integer specifying the number of time points
to simulate.}

\item{sd_step}{A positive numeric value indicating the standard deviation of
the random steps.}

\item{seed}{A positive numeric value setting the seed for random number
generation to ensure reproducibility. If \code{NA} (default), then \code{set.seed()}
is not called at all. If not \code{NA}, then the random number generator state is
reset (to the state before calling this function) upon exiting this function.}
}
\value{
A vector of integers of length \code{n_time_points} with the average
number of occurrences.
}
\description{
This function simulates a timeseries for the average number of occurrences of
a species using a random walk over time.
}
\examples{
simulate_random_walk(
  initial_average_occurrences = 50,
  n_time_points = 10,
  sd_step = 1,
  seed = 123
)
}
\seealso{
Other occurrence: 
\code{\link{create_spatial_pattern}()},
\code{\link{sample_occurrences_from_raster}()},
\code{\link{simulate_timeseries}()}
}
\concept{occurrence}
