% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_simulation_functions.R
\name{map_simulation_functions}
\alias{map_simulation_functions}
\title{Map cube simulation functions over multiple rows of a dataframe}
\usage{
map_simulation_functions(f, df, nested = TRUE)
}
\arguments{
\item{f}{One of five cube simulation functions: \code{simulate_occurrences()},
\code{sample_observations()}, \code{filter_observations()},
\code{add_coordinate_uncertainty()}, or \code{grid_designation()}.}

\item{df}{A dataframe containing multiple rows. Each row is considered a
different species. The columns are function arguments with values used for
mapping \code{f} for each species. \code{df} can have columns that are not used by this
function. They will be retained in the output.}

\item{nested}{Logical. If \code{TRUE} (default), retain list-column containing
dataframes calculated by \code{f}. Otherwise, expand this list-column into rows
and columns.}
}
\value{
In case of \code{nested = TRUE}, a dataframe identical to the input
dataframe \code{df}, but with an extra list-column called \code{mapped_col} containing
an sf object for each row computed by the function specified in \code{f}. In case
of \code{nested = FALSE}, this list-column is expanded into additional rows and
columns.
}
\description{
The function executes a cube simulation function (\code{simulate_occurrences()},
\code{sample_observations()}, \code{filter_observations()},
\code{add_coordinate_uncertainty()}, or \code{grid_designation()}) over multiple rows
of a dataframe containing potentially different function arguments over
multiple columns.
}
\examples{
print("to do")

}
\concept{multispecies_low}
